Changed drbg_external API
Renamed external_drbg to drbg_external for consistency
with other calls.

Changed RANDOM_BYTES_FN length argument from uint64_t to
uint32_t.

+++ test/test_common.c
@@ -6,7 +6,7 @@
 #include "ntru_crypto_drbg.h"
 #include "test_common.h"
 
-uint32_t randombytes(uint8_t *x,uint64_t xlen)
+uint32_t randombytes(uint8_t *x, uint32_t xlen)
 {
   static int fd = -1;
   int i;
